Vertical Tile Sliding
There is a grid of tiles. You can click and drag to drag a single column
When a tile gets out of bounds, it will wrap around to the other side
You can only drag tiles in the interior 5x5. The outer ring of the 7x7
can't be dragged. The outer ting is where the conveyor will be

Bug
Tiles don't wrap around twice. If you wrap them around again, they will be
out of bounds and it you let go of left click, it will set either set
multiple tiles to grid[0] or set them to grid[out_of_bounds] and crash