Towers
Made a tower. It can shoot a bullet every second
A tower spawns on the top and bottom conveyor belts
You can drag the tower off the belt. When you release the mouse button, it
will attach itself to the tile it is over. Then it will follow the tile as
it wraps around

Bug
Doesn't quite work right when you drag a tower over the conveyor